+++ src/main/java/org/betonquest/betonquest/modules/schedule/EventScheduling.java
@@ -26,7 +26,7 @@ public class EventScheduling {
      * Map that contains all types of schedulers,
      * with keys being their names and values holding the scheduler and schedule class.
      */
-    private final Map<String, ScheduleType<?>> scheduleTypes;
+    private final Map<String, ScheduleType<?, ?>> scheduleTypes;
 
     /**
      * Creates a new instance of the event scheduling class.
@@ -34,7 +34,7 @@ public class EventScheduling {
      * @param log           the logger that will be used for logging
      * @param scheduleTypes map containing the schedule types, provided by {@link org.betonquest.betonquest.BetonQuest}
      */
-    public EventScheduling(final BetonQuestLogger log, final Map<String, ScheduleType<?>> scheduleTypes) {
+    public EventScheduling(final BetonQuestLogger log, final Map<String, ScheduleType<?, ?>> scheduleTypes) {
         this.log = log;
         this.scheduleTypes = scheduleTypes;
     }
@@ -66,7 +66,7 @@ public class EventScheduling {
                     );
                     final String type = Optional.ofNullable(scheduleConfig.getString("type"))
                             .orElseThrow(() -> new InstructionParseException("Missing type instruction"));
-                    final ScheduleType<?> scheduleType = Optional.ofNullable(scheduleTypes.get(type))
+                    final ScheduleType<?, ?> scheduleType = Optional.ofNullable(scheduleTypes.get(type))
                             .orElseThrow(() -> new InstructionParseException("The schedule type '" + type + "' is not defined"));
                     scheduleType.createAndScheduleNewInstance(scheduleID, scheduleConfig);
                     log.debug(questPackage, "Parsed schedule '" + scheduleID + "'.");
@@ -89,7 +89,7 @@ public class EventScheduling {
     @SuppressWarnings("PMD.AvoidCatchingGenericException")
     public void startAll() {
         log.debug("Starting schedulers...");
-        for (final ScheduleType<?> type : scheduleTypes.values()) {
+        for (final ScheduleType<?, ?> type : scheduleTypes.values()) {
             try {
                 type.scheduler.start();
             } catch (final Exception e) {
@@ -104,7 +104,7 @@ public class EventScheduling {
     @SuppressWarnings("PMD.AvoidCatchingGenericException")
     public void stopAll() {
         log.debug("Stopping schedulers...");
-        for (final ScheduleType<?> type : scheduleTypes.values()) {
+        for (final ScheduleType<?, ?> type : scheduleTypes.values()) {
             try {
                 type.scheduler.stop();
             } catch (final Exception e) {
@@ -121,7 +121,7 @@ public class EventScheduling {
      * @param <S>           type of the schedule.
      */
     @SuppressWarnings("PMD.PreserveStackTrace")
-    public record ScheduleType<S extends Schedule>(Class<S> scheduleClass, Scheduler<S> scheduler) {
+    public record ScheduleType<S extends Schedule, T>(Class<S> scheduleClass, Scheduler<S, T> scheduler) {
         /* default */ S newScheduleInstance(final ScheduleID scheduleID, final ConfigurationSection scheduleConfig)
                 throws InstructionParseException, InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException {
             try {