+++ src/test/java/org/betonquest/betonquest/api/schedule/SchedulerTest.java
@@ -25,7 +25,7 @@ class SchedulerTest {
 
     @Test
     void testAddSchedule() {
-        final Scheduler<Schedule> scheduler = new MockedScheduler(logger);
+        final Scheduler<Schedule, FictiveTime> scheduler = new MockedScheduler(logger);
         final ScheduleID scheduleID = mock(ScheduleID.class);
         final Schedule schedule = mock(Schedule.class);
         when(schedule.getId()).thenReturn(scheduleID);
@@ -36,7 +36,7 @@ class SchedulerTest {
 
     @Test
     void testStart() {
-        final Scheduler<Schedule> scheduler = new MockedScheduler(logger);
+        final Scheduler<Schedule, FictiveTime> scheduler = new MockedScheduler(logger);
         assertFalse(scheduler.isRunning(), "isRunning should be false before start is called");
         scheduler.start();
         assertTrue(scheduler.isRunning(), "isRunning should be true after start is called");
@@ -45,7 +45,7 @@ class SchedulerTest {
     @Test
     @SuppressWarnings("PMD.JUnitTestContainsTooManyAsserts")
     void testStop() {
-        final Scheduler<Schedule> scheduler = new MockedScheduler(logger);
+        final Scheduler<Schedule, FictiveTime> scheduler = new MockedScheduler(logger);
         final ScheduleID scheduleID = mock(ScheduleID.class);
         final Schedule schedule = mock(Schedule.class);
         scheduler.schedules.put(scheduleID, schedule);
@@ -60,7 +60,7 @@ class SchedulerTest {
     @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
     void testExecuteEvents() {
         try (MockedStatic<BetonQuest> betonQuest = mockStatic(BetonQuest.class)) {
-            final Scheduler<Schedule> scheduler = new MockedScheduler(logger);
+            final Scheduler<Schedule, FictiveTime> scheduler = new MockedScheduler(logger);
             final Schedule schedule = mock(Schedule.class);
             when(schedule.getId()).thenReturn(mock(ScheduleID.class));
             final EventID eventA = mock(EventID.class);
@@ -75,7 +75,7 @@ class SchedulerTest {
     /**
      * Class extending a scheduler without any changes.
      */
-    private static final class MockedScheduler extends Scheduler<Schedule> {
+    private static final class MockedScheduler extends Scheduler<Schedule, FictiveTime> {
         /**
          * Default constructor.
          *
@@ -83,6 +83,11 @@ class SchedulerTest {
          */
         public MockedScheduler(final BetonQuestLogger logger) {
             super(logger);
+        }
+
+        @Override
+        protected FictiveTime getNow() {
+            return new FictiveTime();
         }
     }
 }